% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean}
\alias{clean}
\title{Get Unique Target Classes}
\usage{
clean(data, target)
}
\arguments{
\item{data}{A dataframe}

\item{target}{The target column}
}
\value{
A cleaned dataframe with missing values removed and the categorical target variable converted to a factor
}
\description{
Get Unique Target Classes
}
\examples{
df <- data.frame(Age = c(25, 35, NA, 29), RiskLevel = c("high risk", NA, "low risk", "mid risk"))
clean(df, RiskLevel)
#>   Age    RiskLevel
#> 1   25    high risk
#> 2   29     mid risk

\dontrun{
clean(df, "RiskLevel")
clean(RiskLevel, df)
clean(df["RiskLevel"])
}

}
