% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot Feature Importance from Random Forest}
\usage{
plot_feature_importance(model, output_dir)
}
\arguments{
\item{model}{A trained model object (e.g., Random Forest model)
for extracting feature importance scores.}

\item{output_dir}{Path to save the output .png file}
}
\value{
Saves the feature importance plot and returns the plot object
}
\description{
Plot Feature Importance from Random Forest
}
\examples{
\dontrun{
rf_model <- training_rf_model(train_data)
plot_feature_importance(rf_model, tempdir())
}
}
