% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training_mlr_model}
\alias{training_mlr_model}
\title{Train Multinomial Logistic Regression Model}
\usage{
training_mlr_model(train_data)
}
\arguments{
\item{train_data}{A dataframe containing the training data}
}
\value{
A trained multinomial model
}
\description{
Train Multinomial Logistic Regression Model
}
\examples{
\dontrun{
training_mlr_model(train_data)
}


}
