% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training_rf_model}
\alias{training_rf_model}
\title{Train Random Forest Model}
\usage{
training_rf_model(train_data)
}
\arguments{
\item{train_data}{A dataframe containing the training data}
}
\value{
A trained random forest model
}
\description{
Train Random Forest Model
}
\examples{
\dontrun{
training_rf_model(train_data)
}
}
